Bring :order_with_line_items factory from Solidus 1.2
The order with line items factory was updated in 1.2 to take a stock
location as a transient attribute. This brings that change so we can
specify the stock location which is used for the origin address for packages.

This change is needed to make the integration calculator specs pass in
Solidus 1.0 and 1.1 and can be removed when we no longer want to support those
versions.