Add integration test for FedEx calculators
This change adds integration tests for each of the FedEx calculators for
which we can get rates. There are no tests for any of the `Freight`
calculators, since this extension doesn't seem to support passing of
options to the `carrier.find_rates` method, which would be needed to get
freight rates.

I was not able to get rates for any of the `Saturday` delivery
calculators as well with the addresses I am using for testing. We would
need to investigate further how we can test those.