Use Spree::ActiveShipping::Config in controller
After solidusio/solidus@ec87854 in Solidus core, using a global static
preferences model (rather than a database backed one) is the default.
This means that we should use the global instance of
Spree::ActiveShipping::Config to store and retrieve the preferences. The
new instance of the preferences that was created in the controller was
not persisting the changes since there was no longer a database backing
it.

From now on we should use the global instance of
Spree::ActiveShipping::Config that is assigned in the initializer.